command clean up
