+++ app/Console/Commands/ReportLogging.php
@@ -6,7 +6,7 @@ use App\Scrapping;
 use Carbon\Carbon;
 use Illuminate\Console\Command;
 
-class DailyReportLogging extends Command
+class ReportLogging extends Command
 {
     /**
      * The name and signature of the console command.
@@ -42,6 +42,7 @@ class DailyReportLogging extends Command
         try {
             $scrapping = new Scrapping();
             $year = $this->argument('year') ?? Carbon::now()->year;
+            $this->info('started report sheet scrapping for year '.$year);
             $payment = $scrapping->openTreasury($year)->latest()->initialLogToDatabase();
             $budget_funcCat = $scrapping->openTreasury($year, Scrapping::monthlyBudgetPattern)->initialLogToDatabase();
             $budget_qfuncCat = $scrapping->openTreasury($year, Scrapping::quarterlyBudgetPattern)->initialLogToDatabase();
